Condição de clima

+++ Veiculo/Veiculo/Entities/Relatorio.cs
@@ -6,7 +6,7 @@ namespace Veiculo {
     class Relatorio {
         public Veiculo Veiculo { get; set; }
         public Percurso Percurso { get; set; }
-        public uint KmPercorrida { get; set; }
+        public double KmPercorrida { get; set; }
         public uint QtdAbastecimentos { get; set; }
         public uint QtdCalibragens { get; set; }
         public double LitrosConsumidos { get; set; }
